pyproject.toml
[tool.poetry]
name = "gemini-json-test"
version = "0.1.0"
description = ""
authors = ["ShotaTotsuka <shota.totsuka@blueish.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.5.4"
python-dotenv = "^1.0.1"
google = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
